Move upload function to upload.js
I don't think it's going to be a common utility function.
